refactor: rename token.rs to traits.rs
